# dpanel-api/docs-openapi.yaml
openapi: 3.0.3
info:
  title: DClaw Docs API
  description: API for browsing and searching DClaw app documentation
  version: 1.0.0
servers:
  - url: http://localhost:8088
    description: Local development
paths:
...
  /api/v1/docs/search:
    get:
      summary: Search across all documentation
      operationId: searchDocs
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
          description: Search query
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                type: object
                properties:
                  query:
                    type: string
                  total:
                    type: integer
                  results:
                    type: array
                    items:
                      $ref: "#/components/schemas/SearchResult"
components:
  schemas:
...
    SearchResult:
      type: object
      properties:
        appId:
          type: string
        appName:
          type: string
        path:
          type: string
        title:
          type: string
        snippet:
          type: string